Noted change of handle
I recently changed my GitHub handle from @jdbartlett to @redoPop
and created this account and fork to make sure I don't break any
prior references to my code. The canonical version of this
repository is now:

http://github.com/redoPop/dq

Please use this URL (above) and not the repo you're currently
looking at. This version will not be maintained.

I'm really sorry for any hassle this causes, and I promise to not
make a habit of changing my handle.

❤